TOOL: renames model-transformer folder to internal
