+++ libs/core/util-selection/src/lib/internal/to-array.spec.ts
