+++ libs/core/util-selection/src/lib/internal/to-array.ts
