+++ libs/core/util-selection/src/lib/internal/to-set.ts
