+++ libs/core/util-selection/src/lib/internal/to-states.ts
