+++ libs/core/util-selection/src/lib/internal/types.ts
